refactor: Modal 수정
